Fix spotbugs issue

+++ java/query/src/main/java/sleeper/query/model/output/S3ResultsOutput.java
@@ -89,7 +89,7 @@ public class S3ResultsOutput implements ResultsOutput {
                 }
             }
             LoggedDuration duration = LoggedDuration.between(startTime, Instant.now());
-            double rate = count / duration.getSeconds();
+            double rate = count / (double) duration.getSeconds();
             LOGGER.info("Wrote {} records to {} in {} seconds (rate of {})",
                     count, outputFile, duration, rate);
             return new ResultsOutputInfo(count, Collections.singletonList(outputLocation));
